Changes to flask app structure
I refactored the flask code for my CreateQuiz module. It works very similarly to the original implemenation. I just adjusted it so it could be easier to work with on my end and others' end. It is incomplete but most of the implementation is done. I just need to finish the access portion and the question, answer and choices portion.