Added functionality to flask and some documenation
I finished documenting the most used methods in our project for my module. I do not think some of the methods I created will be used in the final project so I left them out for the meantime.

I also made sure that when the user clicks Finalize Quiz in QuestionCreation.html it does infact save the quiz object with all of the necessary data